write tests to show up the problem
