+++ src/rust/expression/mod.rs
@@ -1,3 +1,4 @@
 pub(crate) mod ast;
+pub(crate) mod eval_ast;
 pub(crate) mod parser;
 pub(crate) mod tokens;
